🚚 rename context files to kebab-case
