+++ packages/kitchn/src/contexts/index.ts
@@ -1,3 +1,3 @@
-export * from "./Combobox";
-export * from "./Theme";
-export * from "./Toasts";
+export * from "./combobox";
+export * from "./theme";
+export * from "./toasts";
