+++ packages/kitchn/src/hooks/use-theme.ts
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { ThemeContext } from "../contexts/Theme";
+import { ThemeContext } from "../contexts/theme";
 
 export const useTheme = () => {
   const context = React.useContext(ThemeContext);
