+++ packages/kitchn/src/native/contexts/index.ts
@@ -1,1 +1,1 @@
-export * from "./Theme";
+export * from "./theme";
